+++ Forge/src/main/java/mod/azure/azurelib/animation/dispatch/command/action/AzAction.java
@@ -6,14 +6,14 @@ import mod.azure.azurelib.animation.dispatch.command.action.codec.AzDispatchActi
 import net.minecraft.util.ResourceLocation;
 
 /**
- * The AzDispatchAction interface serves as a base contract for defining actions that can be dispatched within the
+ * The AzAction interface serves as a base contract for defining actions that can be dispatched within the
  * animation system. It provides methods for handling an action and retrieving its unique resource location identifier.
  * Implementations of this interface encapsulate specific animation-related behaviors, allowing for the modification or
  * control of animation states or properties within an {@link AzAnimator}.
  */
-public interface AzDispatchAction {
+public interface AzAction {
 
-    AzDispatchActionCodec CODEC = new AzDispatchActionCodec();
+    AzActionCodec CODEC = new AzDispatchActionCodec();
 
     void handle(AzDispatchSide originSide, AzAnimator<?> animator);