+++ Forge/src/main/java/mod/azure/azurelib/animation/dispatch/command/action/impl/root/AzRootSetAnimationSpeedAction.java
@@ -3,10 +3,10 @@ package mod.azure.azurelib.animation.dispatch.command.action.impl.root;
 import mod.azure.azurelib.AzureLib;
 import mod.azure.azurelib.animation.AzAnimator;
 import mod.azure.azurelib.animation.dispatch.AzDispatchSide;
-import mod.azure.azurelib.animation.dispatch.command.action.AzDispatchAction;
+import mod.azure.azurelib.animation.dispatch.command.action.AzAction;
 import net.minecraft.util.ResourceLocation;
 
-public class AzRootSetAnimationSpeedAction implements AzDispatchAction {
+public class AzRootSetAnimationSpeedAction implements AzAction {
 
     public double animationSpeed;