+++ Forge/src/main/java/mod/azure/azurelib/animation/dispatch/command/action/impl/root/AzRootSetEasingTypeAction.java
@@ -3,11 +3,11 @@ package mod.azure.azurelib.animation.dispatch.command.action.impl.root;
 import mod.azure.azurelib.AzureLib;
 import mod.azure.azurelib.animation.AzAnimator;
 import mod.azure.azurelib.animation.dispatch.AzDispatchSide;
-import mod.azure.azurelib.animation.dispatch.command.action.AzDispatchAction;
+import mod.azure.azurelib.animation.dispatch.command.action.AzAction;
 import mod.azure.azurelib.animation.easing.AzEasingType;
 import net.minecraft.util.ResourceLocation;
 
-public class AzRootSetEasingTypeAction implements AzDispatchAction {
+public class AzRootSetEasingTypeAction implements AzAction {
     public AzEasingType easingType;
 
     public static final StreamCodec<FriendlyByteBuf, AzRootSetEasingTypeAction> CODEC = StreamCodec.composite(