+++ Forge/src/main/java/mod/azure/azurelib/network/packet/AzBlockEntityDispatchCommandPacket.java
@@ -3,13 +3,13 @@ package mod.azure.azurelib.common.internal.common.network.packet;
 
 import mod.azure.azurelib.animation.AzAnimatorAccessor;
 import mod.azure.azurelib.animation.dispatch.AzDispatchSide;
-import mod.azure.azurelib.animation.dispatch.command.AzDispatchCommand;
+import mod.azure.azurelib.animation.dispatch.command.AzCommand;
 import mod.azure.azurelib.util.ClientUtils;
 import net.minecraft.util.math.BlockPos;
 
 public record AzBlockEntityDispatchCommandPacket(
     BlockPos blockPos,
-    AzDispatchCommand dispatchCommand
+    AzCommand dispatchCommand
 ) implements AbstractPacket {
 
     public static final CustomPacketPayload.Type<AzBlockEntityDispatchCommandPacket> TYPE = new Type<>(
@@ -19,7 +19,7 @@ public record AzBlockEntityDispatchCommandPacket(
     public static final StreamCodec<FriendlyByteBuf, AzBlockEntityDispatchCommandPacket> CODEC = StreamCodec.composite(
         BlockPos.STREAM_CODEC,
         AzBlockEntityDispatchCommandPacket::blockPos,
-        AzDispatchCommand.CODEC,
+        AzCommand.CODEC,
         AzBlockEntityDispatchCommandPacket::dispatchCommand,
         AzBlockEntityDispatchCommandPacket::new
     );