+++ Forge/src/main/java/mod/azure/azurelib/render/armor/AzArmorRendererRegistry.java
@@ -10,17 +10,23 @@ public class AzArmorRendererRegistry {
 
     private static final Map<Item, AzArmorRenderer> ITEM_TO_RENDERER = new HashMap<>();
 
-    private static final Map<Class<? extends Item>, Supplier<AzArmorRenderer>> ITEM_CLASS_TO_RENDERER_SUPPLIER =
-        new HashMap<>();
+    private static final Map<Item, Supplier<AzArmorRenderer>> ITEM_TO_RENDERER_SUPPLIER = new HashMap<>();
 
-    public static void register(Class<? extends Item> itemClass, Supplier<AzArmorRenderer> armorRendererSupplier) {
-        ITEM_CLASS_TO_RENDERER_SUPPLIER.put(itemClass, armorRendererSupplier);
+    public static void register(Item item, Supplier<AzArmorRenderer> armorRendererSupplier) {
+        ITEM_TO_RENDERER_SUPPLIER.put(item, armorRendererSupplier);
+    }
+
+    public static void register(Supplier<AzArmorRenderer> armorRendererSupplier, Item item, Item... items) {
+        register(item, armorRendererSupplier);
+
+        for (Item otherItem : items) {
+            register(otherItem, armorRendererSupplier);
+        }
     }
 
     public static AzArmorRenderer getOrNull(Item item) {
         return ITEM_TO_RENDERER.computeIfAbsent(item, ($) -> {
-            Class itemClass = item.getClass();
-            Supplier<AzArmorRenderer> rendererSupplier = ITEM_CLASS_TO_RENDERER_SUPPLIER.get(itemClass);
+            Supplier<AzArmorRenderer> rendererSupplier = ITEM_TO_RENDERER_SUPPLIER.get(item);
             return rendererSupplier == null ? null : rendererSupplier.get();
         });
     }